Fixes to pipes.c
1) pipes->operationsSemId was removed
2) pipe functions added to syscalls.c and usersyscalls.c
3) changing int to uint32_t in some cases solved overflow problem

+++ Kernel/include/sem.h
@@ -1,13 +1,11 @@
 #ifndef SEM_H
 #define SEM_H
 
-#include <stdlib.h>
-
 int initSemSystem();
-int semOpen(uint32_t id, uint64_t initialValue);
-int semWait(uint32_t id);
-int semPost(uint32_t id);
-int semClose(uint32_t id);
+int semOpen(int id, int initialValue);
+int semWait(int id);
+int semPost(int id);
+int semClose(int id);
 void printSemInfo();
 void acquire(int *lock);
 void release(int *lock);
